% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-class.R
\name{deriv}
\alias{deriv}
\alias{deriv,dual-method}
\alias{deriv,numeric-method}
\title{Extract the derivative (tangent) part of a dual number}
\usage{
deriv(d)

\S4method{deriv}{dual}(d)

\S4method{deriv}{numeric}(d)
}
\arguments{
\item{d}{A \code{dual} object.}
}
\value{
The deriv slot.
}
\description{
Extract the derivative (tangent) part of a dual number
}
\examples{
deriv(dual(3, 1))  # 1
}
