% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-class.R
\name{dual-combine}
\alias{dual-combine}
\alias{c,dualr-method}
\title{Combine dual numbers into a dual_vector}
\usage{
\S4method{c}{dualr}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{A \code{dual} number.}

\item{...}{Additional duals or numerics.}

\item{recursive}{Ignored.}
}
\value{
A \code{dual_vector}.
}
\description{
Combine dual numbers into a dual_vector
}
\examples{
x <- dual_variable(1)
y <- dual_variable(2)
dv <- c(x, y)
length(dv)  # 2

}
