% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-class.R
\name{dual-show}
\alias{dual-show}
\alias{show,dualr-method}
\alias{show,dual_vector-method}
\title{Display a dual number}
\usage{
\S4method{show}{dualr}(object)

\S4method{show}{dual_vector}(object)
}
\arguments{
\item{object}{A \code{dual} object.}
}
\value{
Invisible \code{NULL}; called for side effect of printing.
}
\description{
Display a dual number
}
\examples{
x <- dual(3, 1)
x

dv <- dual_vector(dual(1, 0), dual(2, 1))
dv

}
