% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-class.R
\docType{class}
\name{dual_vector-class}
\alias{dual_vector-class}
\title{Dual Number Vector}
\description{
A container for multiple dual numbers that supports indexing
with \code{[} and \code{[[}, allowing log-likelihood functions to be
written with \code{theta[1]}, \code{theta[2]} notation.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{List of dual objects.}
}}

