% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-helpers.R
\name{hessian}
\alias{hessian}
\title{Compute the Hessian of a log-likelihood}
\usage{
hessian(loglik, theta)
}
\arguments{
\item{loglik}{A function taking a parameter vector and returning a scalar.}

\item{theta}{A numeric vector of parameter values.}
}
\value{
A \code{p x p} numeric matrix (the Hessian).
}
\description{
Uses nested duals (second-order AD) to compute the matrix of second
partial derivatives. Exploits symmetry: only \code{p*(p+1)/2} passes.
}
\examples{
# Normal log-likelihood for mu
ll <- function(theta) {
  mu <- theta[1]
  -0.5 * (2 - mu)^2
}
hessian(ll, 1)  # should be matrix(-1)
}
