% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-class.R
\name{is_dual}
\alias{is_dual}
\title{Test whether an object is a dual number}
\usage{
is_dual(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical.
}
\description{
Test whether an object is a dual number
}
\examples{
is_dual(dual(1, 0))
is_dual(42)
}
