% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-special.R
\name{lbeta}
\alias{lbeta}
\alias{lbeta,numeric,numeric-method}
\alias{lbeta,dual,dual-method}
\alias{lbeta,dual,numeric-method}
\alias{lbeta,numeric,dual-method}
\title{Log-beta function for dual numbers}
\usage{
lbeta(a, b)

\S4method{lbeta}{numeric,numeric}(a, b)

\S4method{lbeta}{dual,dual}(a, b)

\S4method{lbeta}{dual,numeric}(a, b)

\S4method{lbeta}{numeric,dual}(a, b)
}
\arguments{
\item{a}{A numeric or dual value.}

\item{b}{A numeric or dual value.}
}
\value{
\code{lbeta(a, b)} with derivative via digamma.
}
\description{
Log-beta function for dual numbers
}
\examples{
lbeta(2, 3)  # log(beta(2, 3))
a <- dual_variable(2)
value(lbeta(a, 3))  # log(beta(2, 3))
}
