% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-helpers.R
\name{observed_information}
\alias{observed_information}
\title{Compute the observed information matrix}
\usage{
observed_information(loglik, theta)
}
\arguments{
\item{loglik}{A function taking a parameter vector and returning a scalar.}

\item{theta}{A numeric vector of parameter values.}
}
\value{
A \code{p x p} numeric matrix.
}
\description{
Returns the negative Hessian of the log-likelihood at \code{theta}.
}
\examples{
ll <- function(theta) -(theta[1] - 3)^2
observed_information(ll, 2)  # matrix(2)
}
