% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-class.R
\name{value}
\alias{value}
\alias{value,dualr-method}
\alias{value,numeric-method}
\title{Extract the value (primal) part of a dual number}
\usage{
value(d)

\S4method{value}{dualr}(d)

\S4method{value}{numeric}(d)
}
\arguments{
\item{d}{A \code{dual} object.}
}
\value{
The value slot.
}
\description{
Extract the value (primal) part of a dual number
}
\examples{
value(dual(3, 1))  # 3
}
