% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-special.R
\name{beta}
\alias{beta}
\alias{beta,numeric,numeric-method}
\alias{beta,dualr,dualr-method}
\alias{beta,dualr,numeric-method}
\alias{beta,numeric,dualr-method}
\title{Beta function for dual numbers}
\usage{
beta(a, b)

\S4method{beta}{numeric,numeric}(a, b)

\S4method{beta}{dualr,dualr}(a, b)

\S4method{beta}{dualr,numeric}(a, b)

\S4method{beta}{numeric,dualr}(a, b)
}
\arguments{
\item{a}{A numeric or dual value.}

\item{b}{A numeric or dual value.}
}
\value{
\code{beta(a, b)} with derivative.
}
\description{
Computed as \code{exp(lbeta(a, b))}.
}
\examples{
beta(2, 3)  # 1/12
a <- dual_variable(2)
value(beta(a, 3))
}
